% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMEANS.R
\name{KMEANS}
\alias{KMEANS}
\title{K-Means Algorithm}
\usage{
KMEANS(dm, K)
}
\arguments{
\item{dm}{Input data table object}

\item{K}{Number of cluster centers}
}
\value{
A KMEANS object with one value, tot.withinss ( The within cluster sum of
squares value)
}
\description{
K-Means Algorithm
}
\examples{
cluster = KMEANS(ggplot2::mpg, 3)
ss.val = cluster$tot.withinss

}
